+++ tests/test_convert.py
@@ -34,7 +34,7 @@ testdata_dir = Path(os.environ.get('WSIDICOMIZER_TESTDIR', 'tests/testdata'))
 
 
 @pytest.mark.convert
-class ConvertTest(unittest.TestCase):
+class WsiDicomizerConvertTests(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         cls.test_folders = {
@@ -141,7 +141,12 @@ class ConvertTest(unittest.TestCase):
 
     @parameterized.expand(
         [
-            (file_format, file, region, min(file_parameters['include_levels']))
+            (
+                file_format,
+                file,
+                region,
+                file_parameters['lowest_included_pyramid_level']
+            )
             for file_format, format_files in test_parameters.items()
             for file, file_parameters in format_files.items()
             for region in file_parameters['read_region']
@@ -193,7 +198,12 @@ class ConvertTest(unittest.TestCase):
 
     @parameterized.expand(
         [
-            (file_format, file, region, min(file_parameters['include_levels']))
+            (
+                file_format,
+                file,
+                region,
+                file_parameters['lowest_included_pyramid_level']
+            )
             for file_format, format_files in test_parameters.items()
             for file, file_parameters in format_files.items()
             for region in file_parameters['read_region_openslide']