+++ tests/testdata/test_parameters.py
@@ -2,7 +2,8 @@ test_parameters = {
     "svs": {
         "CMU-1/CMU-1.svs": {
             "convert": True,
-            "include_levels": [0, 2, 4],
+            "include_levels": [0, 1, 2],
+            "lowest_included_pyramid_level": 0,
             "photometric_interpretation": "RGB",
             "read_region": [
                 {
@@ -68,7 +69,8 @@ test_parameters = {
         },
         "svs1/input.svs": {
             "convert": True,
-            "include_levels": [0, 2, 4],
+            "include_levels": [0, 1, 2],
+            "lowest_included_pyramid_level": 0,
             "photometric_interpretation": "RGB",
             "read_region": [
                 {
@@ -137,6 +139,7 @@ test_parameters = {
         "czi1/input.czi": {
             "convert": False,
             "include_levels": [0],
+            "lowest_included_pyramid_level": 0,
             "tile_size": 512,
             "photometric_interpretation": "YBR_FULL_422",
             "read_region": [
@@ -161,6 +164,7 @@ test_parameters = {
         "CMU-1/CMU-1.mrxs": {
             "convert": True,
             "include_levels": [4, 6],
+            "lowest_included_pyramid_level": 4,
             "tile_size": 1024,
             "encode_format": 'jpeg2000',
             "encode_quality": 0,
@@ -222,7 +226,8 @@ test_parameters = {
     "ndpi": {
         "CMU-1/CMU-1.ndpi": {
             "convert": True,
-            "include_levels": [4, 6],
+            "include_levels": [2, 3],
+            "lowest_included_pyramid_level": 4,
             "tile_size": 1024,
             "photometric_interpretation": "YBR_FULL_422",
             "read_region": [
@@ -299,7 +304,8 @@ test_parameters = {
         },
         "ndpi1/input.ndpi": {
             "convert": True,
-            "include_levels": [4, 6],
+            "include_levels": [2, 3],
+            "lowest_included_pyramid_level": 4,
             "tile_size": 1024,
             "photometric_interpretation": "YBR_FULL_422",
             "read_region": [
@@ -366,6 +372,7 @@ test_parameters = {
         "ndpi2/input.ndpi": {
             "convert": True,
             "include_levels": [4, 6],
+            "lowest_included_pyramid_level": 4,
             "tile_size": 1024,
             "photometric_interpretation": "YBR_FULL_422",
             "read_region": [
@@ -433,6 +440,7 @@ test_parameters = {
         "philips1/input.tif": {
             "convert": True,
             "include_levels": [4, 5, 6],
+            "lowest_included_pyramid_level": 4,
             "photometric_interpretation": "YBR_FULL_422",
             "read_region": [
                 {